Include our admin section and 'login as other user' feature
I'm using administrate gem to create the admin section. This commit already includes the log in, sign up and log out process.
Also, the feature of "log in as another user" is implemented. Basically we handle session states making possible to an admin user
access the account of another user through the admin section.